% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Balancedness}
\alias{Balancedness}
\title{Calculate balancedness (internal)}
\usage{
Balancedness(netobj, p, q)
}
\arguments{
\item{netobj}{graph network object}

\item{p}{number of assets (nodes)}

\item{q}{number of clusters}
}
\value{
balancedness score
}
\description{
Calculate balancedness (internal)
}
\keyword{internal}
