% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Balancedness_norm}
\alias{Balancedness_norm}
\title{Calculate normalized (0,1) balancedness (internal)}
\usage{
Balancedness_norm(netobj, p, q)
}
\arguments{
\item{netobj}{graph network object}

\item{p}{number of assets (nodes)}

\item{q}{number of clusters}
}
\value{
normalized balancedness score
}
\description{
Calculate normalized (0,1) balancedness (internal)
}
\keyword{internal}
