% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{evaluate_clustering}
\alias{evaluate_clustering}
\title{Evaluate Clustering Performance

Clustering evaluation using multiple metrics}
\usage{
evaluate_clustering(net, true_labels, p, q)
}
\arguments{
\item{net}{graph network object}

\item{true_labels}{True cluster labels (optional)}

\item{p}{number of assets (nodes)}

\item{q}{number of clusters}
}
\value{
Evaluation metrics
}
\description{
Evaluate Clustering Performance

Clustering evaluation using multiple metrics
}
