% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{GINI}
\alias{GINI}
\title{Calculate GINI (internal)}
\usage{
GINI(netobj)
}
\arguments{
\item{netobj}{graph network object}
}
\value{
GINI score
}
\description{
Calculate GINI (internal)
}
\keyword{internal}
