% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn_kcomp_heavytail_TV_graph_online.R
\name{learn_kcomp_heavytail_TV_graph_online}
\alias{learn_kcomp_heavytail_TV_graph_online}
\title{Semi-online inference of a time-varying k-component graph from heavy-tailed data

Computes the Theta matrix of a time-varying graph given the data matrix in a single time frame.}
\usage{
learn_kcomp_heavytail_TV_graph_online(
  X,
  w_lagged = 0,
  sigma_e = exp(1),
  k = 1,
  heavy_type = "gaussian",
  nu = NULL,
  w0 = "naive",
  a0 = 1,
  d = 1,
  gamma = 10,
  eta = 1e-08,
  update_eta = TRUE,
  early_stopping = FALSE,
  rho = 1,
  update_rho = FALSE,
  maxiter = 10000,
  reltol = 1e-05,
  verbose = TRUE,
  record_objective = FALSE
)
}
\arguments{
\item{X}{an T_n x p data matrix, where T_n is the number of observations in a frame (or the frame length) and p is
the number of nodes in the graph.}

\item{sigma_e}{hyperparameter that controls graph weight sparsity and time-consistency}

\item{k}{the number of components of the graph.}

\item{heavy_type}{a string which selects the statistical distribution of the data    .
Valid values are "gaussian" or "student".}

\item{nu}{the degrees of freedom of the Student-t distribution.
Must be a real number greater than 2.}

\item{w0}{initial vector of graph weights. Either a vector of length p(p-1)/2 or
a string indicating the method to compute an initial value.}

\item{a0}{initial value of the VAR coefficient}

\item{d}{the nodes' degrees. Either a vector or a single value.}

\item{gamma}{hyperparameter that controls the sparsity of VAR coefficients in the variations of the weights}

\item{eta}{hyperparameter that controls the effect of the additional regularization to obtain a
k-component graph}

\item{update_eta}{whether to update eta during the optimization.}

\item{rho}{ADMM hyperparameter.}

\item{update_rho}{whether or not to update rho during the optimization.}

\item{maxiter}{maximum number of iterations.}

\item{reltol}{relative tolerance as a convergence criteria.}

\item{verbose}{whether or not to show a progress bar during the iterations.}
}
\value{
A list containing possibly the following elements:
\item{\code{laplacian}}{estimated Laplacian matrix}
\item{\code{adjacency}}{estimated adjacency matrix}
\item{\code{theta}}{estimated Laplacian matrix slack variable}
\item{\code{maxiter}}{number of iterations taken to reach convergence}
\item{\code{convergence}}{boolean flag to indicate whether or not the optimization conv        erged}
\item{\code{eta_seq}}{sequence of values taken by the hyperparameter eta until convergence}
\item{\code{primal_lap_residual}}{primal residual for the Laplacian matrix per iteration}
\item{\code{dual_residual}}{dual residual per iteration}
\item{\code{lagrangian}}{Lagrangian value per iteration}
\item{\code{elapsed_time}}{Time taken to reach convergence}
}
\description{
Semi-online inference of a time-varying k-component graph from heavy-tailed data

Computes the Theta matrix of a time-varying graph given the data matrix in a single time frame.
}
