% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Mode}
\alias{Mode}
\title{Calculate the mode or most frequent element (internal)}
\usage{
Mode(v)
}
\arguments{
\item{v}{input vector}
}
\value{
value of mode
}
\description{
Calculate the mode or most frequent element (internal)
}
\keyword{internal}
