% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_graph.R
\name{plot_graph}
\alias{plot_graph}
\title{Plot a clustering graph from adjacency matrix

Plots a clustering graph based on a given adjacency matrix and the ground-truth labels for the clusters.}
\usage{
plot_graph(
  adjacency,
  true_labels,
  node_names = NULL,
  implied_clusters = FALSE,
  verbose = FALSE,
  Coords = NULL
)
}
\arguments{
\item{adjacency}{p by p adjacency matrix.}

\item{true_labels}{the numeric vector of true labels (of length p)}

\item{node_names}{the vector of names of the nodes, e.g., stocks (of length p)}

\item{implied_clusters}{whether to color the nodes based on implied clusters or not}

\item{Coords}{input graph node coordinates}
}
\value{
The network object:
\item{\code{graph_net}}{the graph network object that can be used for plotting}
\item{\code{metric}}{the results of the clustering measures}
\item{\code{Coords}}{output graph node coordinates}
}
\description{
Plot a clustering graph from adjacency matrix

Plots a clustering graph based on a given adjacency matrix and the ground-truth labels for the clusters.
}
